clrcore: replace RCW creation with custom FastMethod COM invocation to bypass Mono domain bug
Mono seems to reuse the RCW JIT methods for IScriptHost across app domains, but when doing exception unwinding it wants all stack frames up to a domain transition to be part of the *current* app domain.

This would lead to a weird kind of crash. After hours of debugging, it was decided the best course of action was, in fact, to not use RCW/COM interop at all in this context, and instead directly invoke the method pointers using DynamicMethod.