feature: deterministic interior hashes over the network to remove the need for interior proxy ordering files
This is currently not very performance-optimized, it should be seen as an early proof-of-concept. Toggled using debug builds or having `new_interior_hash` global policy, this will be toggled globally later on once the parser changes propagate.

If anyone wants to do limited testing (either on canary - requires all clients to have canary! - or when there's a prod push), this policy might be manually grantable.