fix: untested and probably misguided attempt at resolving object ID stealing issues
ReassignEntity was implemented for non-clients without the realization that removing an entity from its owner would lead to the owner returning the object ID to its own pool.

This changeset reorders removal around so that the reassignment will only take place once the owner has *for sure* been notified of the entity's relevancy removal, and implements a state where the object ID won't be returned to the owner pool.