tweak(client/launcher): speed up file verification
CheckFileOutdatedWithUI was spending a lot of CPU time in a) updating the progress bar and b) a really weird SHA1 library.

Instead of those, we're only updating the progress bar on a 0.2% increment, and since we're linking OpenSSL anyway, replace the SHA1 library with OpenSSL's SHA1 functions.

After these changes, we're apparently verifying `update.rpf` at ~380 MB/s on a NVMe disk.