fix(streaming): return low LSN for pgRawStreamer
This - and not the flag in `rage::strPackfileManager::ms_files` - seems to be the controlling flag to move pgRawStreamer to use the other streaming thread, and therefore unblock streaming from the local disk while network streaming is in progress.

Thanks for bearing with us while we tried to figure this one out!