fix(onesync/client): calibrate distance culling range for distant scene update
At the expense of reducing its effect on bigmode, this range should match the usual range used on the server for population-spawned entities so that vehicles will not be in an inconsistent state when migrating (e.g. affected by physics but not re-blended from the network sync node, leading to near-fallen helicopters).