fix(onesync/server): infinite recursion
A ped can not be its own vehicle.

A ped can not be its own vehicle.

A ped can not be its own vehicle.

A ped can not be its own vehicle.

A ped can not be its own vehicle.

A ped can not be its own vehicle.

A ped can not be its own vehicle.

A ped can not be its own vehicle.

A ped can not be its own vehicle.

A ped can not be its own vehicle.

A ped can not be its own vehicle.

A ped can not be its own vehicle.

A ped can not be its own vehicle.

A ped can not be its own vehicle.

A ped can not be its own vehicle.

A ped can not be its own vehicle.