fix(mumble/five): sound allocation fixes
We now create Mumble sounds in dedicated allocation buckets. Every allocation bucket has a size of 64 (changing this crashes the game), a single audExternalStreamSound takes 2 entries, and we have 4 buckets for a total of 128 sounds at most.

In addition to that, implement a few sound de/reallocation fixes to prevent certain scenarios such as sink deletion from leaking excessive sounds.