Added initial support for Game Events (Emitting and Reacting)
This update will re-enable the hooking for the missing game events and retrieves all entities associated with it.

I've created separate scripting events that are being used, this, to allow fitting arguments to be passed to scripts and not to break existing listeners. They are called "gameEventEmit" and "gameEventReact" respectively as there are 2 version and they both act slightly different from each other.

Events like CEventShocking*, CEventPotentialWalkIntoVehicle, and CEventPotentialGetRunOver already retrieve the world position.
To support more events' data retrieval we would need to analyze the event data and setup proper conversions.

Adding include dir: "..\..\..\vendor\msgpack-c\include" to 2 projects: gta-streaming-five and handling-loader-five.