fix(rage-device/five): Error out on cloud files
The issue:
The application seems to get stuck in a loop when cloud files are
accessed, this is caused by users installing their GTA on the cloud,
causing the game to neither load, nor 'gracefully' exit.

The solution:
To correct this, we simply get the file attributes and check if
the following flag is set: `FILE_ATTRIBUTE_RECALL_ON_DATA_ACCESS`. If
it's set, we return an invalid handle, so it's passed down the stack.
Once this errors, `FatalError` under `OpenArchiveWrap` will be
triggered. I have modified the message so any affected clients are
notified in the event this occurs.