tweak(net/five): support clipset no-op for 2060+
This code was commented out on b2060 and above because of the caller
being obfuscated.

However, this function was only used in this one single location, and
could easily be patched out from that perspective.

Since the function is a lot like EndNetworkSession, the pattern is in
one of the parts that *do* differ.

In addition to that, to support switching from SP -> MP without breaking
the clip set manager, we check the network session state and call the
original instead of blatantly no-op'ing the function.