fix(launcher/minidump): bundle dbghelp.dll for downlevel
Symbolication didn't work for very-downlevel systems as the dbghelp.dll
used on these systems was sufficiently old to not match the SDK headers.

Given the required dependency DLLs, and some code to use this dbghelp
for symbolication (using it elsewhere as well led to crash handling
failing entirely), we now have working symbolication even on Windows 7.