tweak(gta/five): correct LoadingScreenContext for network-kill
The last two fixes for clearing out the streamer to reduce this hang
type did not have any effect.

However, it appears there's a loading screen context type ('INSTALL')
which doesn't try to do any blocking load and will just set the loading
screen render thread override (and is faster than old LAST_FRAME).

This still won't fix the fact the game is 'stuck' and won't correctly
reload, but at least this is better than the game freezing, followed by
a CEF crash and/or a watchdog timeout.