tweak(launcher): increase max stack size (for v8)
V8 expects 1 MB of stack available for itself, discounting any code out
or inside. As such, a stack overflow in V8 scripts would usually lead to
a native stack overflow crash and not the V8 runtime itself trapping on
its internal stack check.

Chromium source code shows 1.5 MB as a safe assumption, even though we
are on 64-bit here and therefore have 8b-sized stack slots, it's not
likely that this limit will be badly hit (and even if, it will still
crash).

This change may have impact on memory usage for background threads and
would need some checkup at a later time.