fix(client/launcher): no /STACK for main .exe
The last commit (5376bb10d6fc94ff41ba628be558c1a859c296c2) was found to
also have the stack override for the main FiveM.exe in `release` builds.

This was apparently due to a confusing `filter {}` statement in the
project file.

Even though this won't 'prevent' an unneeded launcher bootstrap even on
beta/prod, we should correct this.