fix(net/five): broken settings restart in net since 59b493ae67175d69491d1b868d6f122e56aa98dc
The fix commit mentioned in the subject *incorrectly* assumed there were
two places to patch, and that return_function on this function would
correctly 'suppress' the network disconnect.

This is incorrect: the 2nd Arxan-obfuscated caller is actually a native,
and overwriting with return_function won't execute the `else if` used
when the game is trying to exit/restart due to a setting change.

The second patch from d4f8d616c848be7b345144ff6924049a0eb8e419 simply
does not exist anymore in newer builds, nor was it ever required (as the
patches were tested on b1604+, each of these referencing broken replay
editor code for the patch).