fix(data/client): update Mono dependency
This fixes a bug where the Mono IL verifier was skipped for some non-
platform assemblies.

Compatibility *may* be affected if any users were relying on assembly
verification being skipped, but it should be noted that this was not an
intended scenario, and the risks of keeping the verifier disabled/broken
are too high to justify doing so, even for compatibility.