[DEV-731] Allows use of escrowed scripts in non-escrowed resources
When a non-escrowed resource includes an escrowed script, the "before
load" hook ends up being called on the non-escrowed resource, and not on
the escrowed resource. As a result, the Lua bytecode is never decrypted
and throws an error.

This resolves that issue by doing the same type of path parsing
used elsewhere (`@resource/file`) to call the proper resource rather
than the parent resource (in this case, non-escrowed).

Also worth noting that any resource that includes an escrowed script
will also need to enable Lua 5.4, as escrowed scripts require Lua 5.4.