fix(gamestate/five): dummy object stub ABI
This broke as of 0510654472b28c3cce0f3edb31974d4b5fccb89c, as reported
only recently in #1452 as a *reproducible* case of 'blowing up a gas
pump leads to the pump instantly being recreated'.

It turns out that the initial fix made a few mistakes in the ASM stub:

    1. `al` was overwritten with the post-restore `al`, and
    2. `rax` was saved in the scratch space for the callee, so it would
       often contain nonsense, causing 'Player Wants Control' to always
       be true.

Fixes #1452 - at least for the reproducible regression case, hopefully
for all cases however.