fix(game/five): mitigate tools compressing "script_rt" textures
Some tools (among which, a 'ToolkitV') purporting to 'optimize texture
dictionaries' unconditionally compress any uncompressed texture they
find in a texture dictionary.

However, in doing so, they ignore game code that handles any 'script_rt'
texture as if it were uncompressed, and therefore the initial data in
such a texture may lead to an out-of-bounds read in the GPU UMD when the
texture is created.

We can trivially detect this scenario, and uncompress the texture + set
a correct uncompressed format.

This fixes #1447.