tweak(net): don't request a Steam ticket if not needed
Even if the server doesn't/can't accept a Steam authentication ticket,
client code will still request such a ticket if Steam is running, which
may fail if Steam is down or otherwise 'broken'.

As of this change, the server will advertise (before connecting, since
initConnect is still one-shot) whether it expects a Steam ticket, and if
a server explicitly says it doesn't, the client will not try to request
and send such a ticket.

This was requested by Flawws a while ago.