fix(ext/cfx-ui): filter help box styling tweaks
Even though old cfx-ui is getting deprecated soon, people asked about
this help box on the forums (https://forum.cfx.re/t/4862609) where it
looked as if the user didn't understand the `~` due to the font used by
UI nowadays having a very unclear tilde.

As such, we change the font for the 'example' to be monospaced, and also
fix the white space next to the ` - ` being ignored.

Both are likely styling oversights when the 2020.2 theme was initially
implemented.