tweak(client): reduce amount of needed ros_* files (attempt 2)
We don't actually load SCUI in the bundled CEF from the SC SDK, so it is
a bit redundant to download ~50 MB of additional files to support this.

Instead, we will return the correct versions (and presence) from a hook
to satisfy the file corruption checks in the main SC SDK library.

This is a respin of 6d1d43156718c2cad9533a8fab9f3c6d600dfb86 but with a
fix for RedM failing to initialize.