tweak(graphics/rdr3): don't ignore pipeline cache
It turns out that the game, if not 'cleanly exited' (which is rare) will
invalidate the SGA pipeline cache file, leading to a very slow launch
especially on NVIDIA GPUs since we were forced to disable the NVIDIA UMD
shader compiler cache.

Instead, we'll just ignore this, but in the future, this might check for
a GPU-related failure in last run.