fix(scripting/mono): funcref regression in arg checks
As of 89b309452948a458a53dc965148ab3dc08331421, function references did
not work anymore, as the code (RemoteFunctionReference.InvokeNative)
used there forced a `char*` declaration to be cast to `IntPtr`.

This native should be refactored sometime to use ResultAsObject, but for
now we're special-casing it in the Mono ScRT.

Fixes #1488, reluctantly. Less abuse with random '+1' emoji responses or
a nasty (since-deleted) comment asking '3 Likes but no answers?' for
some reason.