tweak(gamestate/server): faster sync data parsing
These changes have been sitting around my local working tree for a few
months now, and seem to work fine.

Intent here was to specialize code for create/sync, so that there'd be
less branching in this loop. Eventually this was meant to also refactor
Unparse, but this never happened.