fix(scripting/lua): check for non-callable event handler
A user reported a vague error at some later point when having registered
an event using the (wrong) syntax `RegisterNetEvent('ev', true)` where
once the event got invoked, the handler (boolean of 'true') would be
called/getinfo'd and that would then error out of the event handler
function.

Ideally we'd report this on registration, but that'd require the warning
logic to be implemented (see #888), as an error would potentially break
existing scripts.