tweak(debug/net): netgraph imgui window flags
A local change already had the focus flag here (likely from diagnosing
console focus), and then #1551 mentioned this may show a scrollbar as
well.

Since this is a bit too trivial a change to then deal with weird merge
madness for local worktree changes, that change is taken in here too.

Closes #1551.