fix(scripting/lua): order for CreateThread from a thread
Since the 'bookmarks' change, Wait(n) in a thread would reschedule the
thread *after* any CreateThread usage in the same bookmark.

This is a bit of a nasty hack (ideally, the script host should put the
bookmark back where it came from) and it doesn't work on CreateThreadNow
but it is enough to fix the draw order in `kgv-phone`.