fix(vfs): deadlock-inducing mutex usage in StreamDevice
A Sentry hang showed a deadlock between the resource manager mutex and
the mutex on the compcache:/ device, since the inner function for a new
handle would execute with the vfs::StreamDevice lock held.

We fix this by using the same approach as other AllocateHandle funcs,
which is to instantly set valid=true, and also switch around opening an
underlying stream and allocating a handle since AllocateHandle does not
fail anymore.