tweak(device/five): fiDeviceRelative length methods perf
These methods would normally just be the base fiDevice's implementation,
which if the underlying device is a fiPackfile and it's checking a zlib
file, it'd lead to the file being decompressed (or worse, decrypted) two
times, instead of just returning the length from the packfile header.

Instead of that, we patch it with a function that'll just forward these
calls to the underlying device.