tweak(client): handle RtlReportException crashes
Some exceptions (such as STATUS_HEAP_CORRUPTION) get passed through the
RtlReportFatalFailure handler, which has an exception handler that will
call NtTerminateProcess right away after submitting to WER, skipping the
RtlDispatchException fallback handler and leading to a 'silent' game
exit instead of invoking the minidump handler.

It appears that we however can hook RtlReportException and call our own
handler in there like we do in RtlDispatchException, and as such heap
corruption at least shows the custom error and submits a Sentry report.

(we, of course, lack the heap-specific data that'd be submitted to WER
 here and the stack is a little unfriendly like this, but that is a
 later concern in case this messes up crash analysis)

This is just the *client* bit of a fix - FXServer would have this issue
equally so on Windows, which apparently came up in confusion regarding
the issue fixed in 172d9bf8e98506d2f6e0ae6a67e3de9fdaac86a2 as a common
exception there is the exact same heap corruption, but at this time the
server launcher doesn't expose the same exports the client launcher does
and SEHTableHandler is very much client-specific at this time.