fix(nui/core): D3D init deadlock with in-process GPU
Introduced in aaf3a40108feb5aaa759f2e90e114e9ef57e0615, this would hit
a deadlock if this code ran with the Flush mutex held (such as when the
NUI in-process GPU option is enabled) as shared_mutex does not directly
have recursive ownership support.

Thanks to PichotM for the report + memory dump!