fix(client): nvidia shader cache workaround race condition
Even the DRS APIs have a race condition that at times leads to profiles
being saved that for whatever reason don't contain the shader disk cache
override.

This wasn't reproducible, but reworking this function (also adding logs
and renaming the profile to be more visible in NVIDIA profile viewers)
to add a mutex may help here.