tweak(net/http-server): flag to remove EASTL usage
The known 'broken' build range from https://forum.cfx.re/t/4975484 looks
like it includes an EASTL bump commit (2ddeb1559) as likely culprit.

Since the crash diagnosis also includes the `"200"` string (as in H2's
:status field) being written in spurious places, the use of EASTL might
be suspect here.

This is done using an `#if` so it can easily be changed when trying to
work on issues caused by this.

The flag is also toggled off for now to make this a valid build to test
for the issue there.