fix(graphics/rdr3): draws being cut off on D3D12 with frame scaling
Code here used to set a depth/stencil buffer to the base DepthBuffer,
but on D3D12 this would then clip the entire viewport to the size of the
depth buffer instead of the viewport/render target, which may mismatch
if frame scaling is enabled.

Apparently, this can pass 'null' as well, which fixes this discrepancy.