tweak(client): redirect 'Documents\Rockstar Games' to AppData\Local
A common case of 'stuck' game processes and the subsequent 'could not
lock file' error seems to be resolved by users deleting the Rockstar dir
in their Documents.

It could be a common cloud file provider somehow causes a weird kernel
hang of this kind, but to circumvent this we just move all file access
there to AppData\Local.

This might affect 'User Music' placement so it *may* need end-user docs
for this.