fix(ros): sessionKey offset for RGL signin
SC SDK 2.2.6.9 moved the sessionKey field down a bit, for the first time in a long time.

This'd break automatic signin from the RGL in an unusual way (sessionKey would be read as a buffer of NULLs, leading to requests failing).

Ideally this'd be paired with an assertion or obtained from a pattern scan somewhere, but for this commit, these offsets have just been factored out in `constexpr`s.