feat(net/client): Add state bag methods
The modifications in this pull request introduce two new natives along with supporting methods for state bags.

STATE_BAG_HAS_VALUE: This native is designed to eliminate unnecessary deserialization of state bag values. It serves the purpose of determining whether a value exists for a specific key within the state bag without the need for full deserialization.

GET_STATE_BAG_KEYS: This native expands the utility of state bags by enabling the construction of custom models around the state bag without the necessity to manually check each key for existence. This enhancement streamlines the process of working with state bags and facilitates the creation of more efficient and organized models.