tweak(server): prevent packet routing in onesync and double vector allocation
onesync does not use the packet routing. It only sends the sync data directly to the server. By moving the RoutingPacketHandler to a extra file its possible to run automated tests against it. Use remaining bytes instead of packet length for RoutingPacketHandler  payload size to remove these extra bytes in a future net version and to reduce complexity of size checks.