tweak(server): move rcon, improve GetStatusOutOfBand
The RconOutOfBand packet handler is now a standalone file. The out of band handlers are now initialized with the server when used the first time. The GetStatusOutOfBand handler is faster and has a ConVar sv_returnClientsListInGetStatus to toggle if the client list should be returned, this is by default enabled.