fix(game/five): display correct amount of dlcs that will be loaded
Previously number of dlcs and extra packages was hardcoded per game
version. This led to messages like "Invoking ... (62 out of 55)" on
loading screen.

There were also some off by one errors which led to counter being
displayed out of order (e.g. first "(57 out of 55)" and right after it
"15 out of 55").

The updated system dynamically checks the number of dlcs that will be
loaded. So the counter is correct regardless of the game version and
doesn't need to be set manually anymore.