feat(shared): tcp connection timeout
The tcp connections now have a timeout that will close them when no data is send to send or received from them for a specific amount of time. A running http handler will increase the timeout to not close a connection when the server is computing a http response.

Add sv_httpHandlerConnectionTimeoutSeconds:

This ConVar defines the amount of seconds to increase the timeout of a tcp connection while a http handler is executed. This is by default set to 5 minutes. When the http handler is finished the time will be set back to the default tcp connection timeout.

Add sv_tcpConnectionTimeoutSeconds:

This ConVar defines the amount of seconds to wait for a tcp connection before closing it when no data is received or send. This is by default set to 5 seconds.