fix(game/five): fix switching to build 1604
After we moved list of game builds into lua dictionary in
premake5_builds.lua file - we can not rely on order of builds in
GAME_BUILDS define. However GetGameBuildInit was doing so. It would set
the default build to be the last one in the list. Which was expected to
be the lowest one.

On the other hand, when switching builds, we dont write build 1604 as a
flag in CLI because we expect it to be the default one anyway.

As a result - when trying to writch into build 1604 game would go into
infinite restart loop.

This commit sets default build in GetGameBuildInit explicitly. We so it
the same way in all other places, so there is no reason to try to get
smallest build from GAME_BUILDS list.