feat(data): Add fwPortalSceneGraphNode pool to gameconfig
Currently by default, fwPortalSceneGraphNode is set to 400, servers with many interiors and in turn lots of portals, this can cause a game crash. its possible to simply add this pool to the config to increase it.

currently it says "Unknown Pool full - 400", which makes sense because fwPortalSceneGraphNode is missing from the config therefore its unknown.