feat(game/five): Add natives to operate with drawables through collections
Added analogues to natives like SET_PED_COMPONENT_VARIATION that use
collection name and local (relative to the start of collection) index
instead of the global index.

With the old natives indexes would often shift after Title update which
would make migration to the newer game build harder. With the collection
relative indexing this would not be a problem as the local indexes
remain stable.

The new natives basically provide interface to the base game
functionality that already existed but was unavailable before.