feature(game): allow server to increase pool sizes
Sizes of pools are increased with "increase_pool_size" command that
accepts name of the pool and number of elelments by which the pool size
should be increased.

Which pools can be resized and by how much is limited by the cfx server.
The limitations are checked on server and client side.

When server requests differen pool sizes than the ones the client
already has - the game will be restarted. The same mechanism as for the
build switch is used.