fix(game/five): prevent memory corruption in gta-streaming-five
Before build 3258 we were using rage::fwEntity::SetModelId to get
archetype by model index, because it would not do anything besides
setting the archetype. However on 3258+ this function is no longer
available, so we used another one. However the new function expects an
input of higher size and may have some side efects if the input is not
initialized properly. We forgot to increase the fakeEntity size which
led to memory corruption.

The memory corruption caused a memory leak in fwArchetypePooledMap pool
and a few other pools (neither of which are not tracked by Pool Monitor
unfortunately). As a result we would run out of space in those pools and
some entities would fail to be created. The way visibility for lods is
set is that the lod is displayed until all its children (higher
resolution pieces of image) are created and added to lod hirerarchy.
Because otherwise we may have holes in textures. So when one or two
children of a lod would fail to be created (usually a bush or something
insignificant) - the lod would stay visible. I.e. a blurred texture
would be drawn over the high resolution one. Which was reported in issue
https://github.com/citizenfx/fivem/issues/2752.

Exactly the same symptoms may be visible if map is just overloaded with
resources and the archetype pools run out of space for natural reasons.
Unfortunately there fix there is more complex than just increasing a
number because the current size limit is already very close to the data
type size limit (64000 vs 2^16). It should be increased in followup
commits nevetheless.

Some people mentioned they've seen this issue on builds before 3258 (but
much less often). It could be due to too much resources (as described
above), or maybe there is some other (much smaller) issue as well. But I
believe that the cause for blurred lods to stay visible is the same -
some of it's children are not created.