feat(game/five): introduce sv_replaceExeToSwitchBuilds feature flag
When set to false, this flag indicates that instead of loading the old
game build executable and corresponding files - we should use the latest
stable game build executable and only replace certain parts of
update.rpf. By default the flag is set to true which corresponds to the
default behavior as it was before this commit.

Also introduce sv_enforceGameBuild 1, which corresponds to the game
version without any DLC content. If build 1 is set - the game will run
as if sv_replaceExeToSwitchBuilds is set to false.

In the future this will allow to deprecate old executables while
still allowing servers to control set of content/dlcs that they want to
have. Deprecating old executables will significatly simplify our
codebase and speed up development.