fix(game/five): fix video editor crash when loading clips
Video editor would sometimes crash during clip loading. This happened
when editing clips that were recorded on servers with different
resources or when opening the video editor after connecting to a server
with custom resources.

The crash happen because we didn't clean up the streaming files
properly before opening a new clip. As a result some entities would
point to wrong archetypes.

Apply the same cleanup logic for loading a clip as we do on game
disconnect.