fix(game/five): correctly unregister duplicated archetypes
In case of duplicate archetypes in different fwMapTypes assets - the
game will not unregister streaming archetypes correctly. This leads to a
memory leak in the archetype pool and, eventually, visual bugs and
crash.

This is a rewrite for a fix that was attempted five years ago. But seems
like it never actually worked.