fix(game/five): set correct server side game build on new build system
This is a followup fix for:
https://github.com/citizenfx/fivem/commit/31413e95ed891a9fd6392f6b8f19d8f7fece1632

In ServerGameState we were relying on GTA_FIVE and IS_RDR3 macros which
are not defined on the server side. This commit fixes that depencency.

I have not encountered any issues caused by this myself, but it could
potentially break things in some unexpected places.