feat(scripting): optimize native invocation
Optimized logic relating to native invocation, and separated the argument and result buffers.

This includes a number of optimizations of how natives are invoked, including:
* Avoiding IScriptHost::InvokeNative, which removes multiple levels of indirection, and allows better caching of the type info and handler.
* Improving inlining, and reducing function calls.
* Reducing branching/overhead, particularly when pushing Lua arguments.
* Returning a closure in Lua_GetNativeHandler, removing the need for Lua_InvokeNative2

Separating the argument and result buffers also removes the ability to treat arguments as results,
improving native security, and removing the overhead of having to check for these issues.