fix(streaming/five): improve archetype handling
Although the logic in 0943327f22779f44dc7f4f1bef37103a801b26ad was likely correct, it didn't address a separate issue, which is that LookupModelId retrieves an archetype's model id though the ArchetypeMap.

When there are duplicate archetypes, this breaks stuff, because that model id doesn't necessarily correspond to the provided archetype.
Instead, just cache the model id inside of the archetype and return that.

Also, protect archetypes used by scripted objects, to ensure they are not prematurely unloaded.