fix(game/five): update gameconfig.xml for 3258 and 3323
In 3258 config we set mission thread stack size to 60500 as it was in
previous build. But the game actually expects it to be 62500. Because of
this story mode cinematics in missions didn't load. Which would lead
either to infinite game loading or to NPC getting frozen.

For 3323 we didn't have CGameConfig at all. Which let to story mode not
being able to load at all.

The 3323 config changes are taken from base game update.rpf/common/data
files.