fix(game/five): return dlc count check to 1604 when using new build system
Usually when running on 1604 we check that there are exactly 132 DLCs
mounted. However when running with sv_replaceExeToSwitchBuilds feature
flag this check always gets skipped. Even though the number of loaded
DLCs is still relevant. This would let some broken client to connect to
a server and experience unplayable game. Fix the check.