feat(game/five): allow increasing ms_entityDescPool
Add support for increasing ms_entityDescPool through the regular
increase_pool_size server commans. The name of the pool is
"EntityDescPool". The current max increase is 20480 (may be adjusted in
the future).

To increase the pool add following (with preferred size increase) to
your server config: `increase_pool_size "EntityDescPool" 5000`.

The increase is needed if your clients experience crashes in
rage::fwBaseEntityContainer::GrowDataHandle. Usually with the following
error:
```
RAGE error: ERR_GEN_ENT_CONT_3
A game error (at 000000014165ae44) caused the game to stop working.

Game memory error. Please reboot and restart the game.
```

EntityDescPool is not added to pool monitor because it doesn't store the
number of elements in it explicitly. Instead it stores list of gaps/free
spaces. We could add multiple patches to keep track of number of
entities without adding much performance load. But I'm not sure if the
added complexity and maintenance cost worth the benefits of having
EntityDescPool in the pool monitor. We can reconsider this decision
later if needed.