fix(game/five): remove delay on remote client when firing a weapon
Fix for https://github.com/citizenfx/fivem/issues/2934

On build 3258+ there would be a significant delay between shot in source
client and shot in remote client. This would only appear when aiming on
foot. E.g. firing a weapon from a car worked fine.

The problem appeared because for some reason in 3258+ GTA5 started
syncronysing `CWeapon::m_State` ower the network. It is not the case in
3095 and below. The state is now updated in
`CNetObjPed::UpdateWeaponForPed()`. This leads to `m_State` be set to
`STATE_WAITING_TO_FIRE` on remote client before the shot actually
happens on remote client. So `CTaskAimGunOnFoot::Aiming_OnUpdate()` is
blocked until `CWeapon::m_State` is synchronised back to `STATE_READY`
and only after that shot on remote client actually happens.

This patch returns the original 3095 behavior.

It is unclear why the base game behavior was changed this way and why it
doesn't cause issues in GTA Online. My assumption is that it has
something to do with the order of events: e.g. maybe we somehow
synchronise states too early. And the state sync was added just in case.
Regardless of the reasons - I've tested build 3258+ with the patch and
there seems to be no issues, synchronisation works fine, no new problems
appear because of it (both with onesync on and off). The `CWeapon` state
is being updated as intended by the usual `CWeapon::ProcessState` (same
as it does on 3095)