fix(game/five): sync ped inventory on update
In the base game the CPedInventoryDataNode is defined as MIGRATE_NODE.
I.e. is it synced on creation and when owner is changed. However in
onesync we don't support the migration event currently. This leads to
inventory data being lost when ped owner is changed. E.g. peds loose
their weapon when players moves too far away and then returns.

Update the inventory node synchronisation flags to make it sync on
update event as well. One sync caches the data, so the inventory data is
sent over the network only when there is a change.