feat(gta-streaming-five): allow to edit scene graph pools
tweak(gta-streaming-five):

Change the initPoolsLocation at the top of the function.
Index the pools by eSceneGraphPool instead raw index
Change memory offsets in other functions(computeSceneNode, assignScratch, clearVisData)

tweak(gta-streaming-five): patch FIRST_ROOM_SCENE_NODE_INDEX

Changed the first pattern to the begin of the function.
Added the GetGbufScreenQuadPair function patch
Changed the value of the FIRST_ROOM_SCENE_NODE_INDEX(FW_EXTERIOR_SCENE_GRAPH_NODE + FW_STREAMED_SCENE_GRAPH_NODE + FW_INTERIOR_SCENE_GRAPH_NODE - 1)

feat(gta-streaming-five): patch sky portal flags

As Iridium said the at `83 B9 ? ? ? ? ? 4C 8B CA 4C 8B C1` function needed to be patched because the size of the skyPortalFlags was the size of the FW_PORTAL_SCENE_GRAPH_NODE pool. Since we doubled it(800) that size was not updated.
I created a function that returns the address of the new skyPortalFlag with its correct size based on the address of the class self, so if there is another instance of the same class it will have a different skyPortalFlags.
I patched the AddSkyPortal function to use the new SkyPortalFlag array and the same with the SetIsSkyPortalVisInterior.
Fixed the offset of the pool size constructor of FW_FIXED_ENTITY_CONTAINER.

tweak(gta-streaming-five): rewrite AddSkyPortal function and finish SetIsSkyPortalVisInterior patch

fix(gta-streaming-five): finish SetIsSkyPortalVisInterior patch

Changed r8 register to match the original(r8d)
Changed the value that take r8d to `(pools[eSceneGraphPool::FW_PORTAL_SCENE_GRAPH_NODE].poolSize + 7) / 8`
Add original missing code `mov(dword_ptr[rcx+0x378], r9d)`

fix(gta-streaming-five): SetIsSkyPortalVisInterior with wrong registers values

Previously I was storing the address of `skyPortalFlagArray` in rcx, which was not correct because later in `mov(dword_ptr[rcx+0x378], r9d);` rcx was used to set `m_skyPortalCount` to 0, since having the address of skyPortalFlagArray as rcx this was not working correctly. Now we are storing the value of the original code in rcx.
I have removed `add(rcx, 0x37C);` because it is the offset to access the `skyPortalFlagArray` of the class but since we are allocating it in the c++ code, the only thing that offset was doing was jumping over a memory hole that was not used.
Finally after setting `m_skyPortalCount` to 0 the value of skyPortalFlagArray is stored in rcx to later call `memset`.

tweak(gta-streaming-five): rework AddSkyPortal & validate skyPortalCount

Refactor the `AddSkyPortal` function.
Sometimes `skyPortalCount` may take the value of -1, that produce a crash later when `GetStoredScreenQuadIndex` return a invalid index.